+++ app/src/main/java/org/sopt/certi/presentation/ui/setting/SettingScreen.kt
@@ -49,7 +49,6 @@ fun SettingRoute(
         viewModel.sideEffect.flowWithLifecycle(lifecycleOwner.lifecycle).collect {
             when (it) {
                 SettingSideEffect.NavigateToLogin -> navigateToLogin()
-                else -> {}
             }
         }
     }
