+++ app/src/main/java/org/sopt/certi/presentation/ui/setting/SettingViewModel.kt
@@ -11,9 +11,6 @@ import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 import org.sopt.certi.core.network.TokenManager
 import org.sopt.certi.domain.usecase.auth.WithDrawUseCase
-import org.sopt.certi.domain.usecase.user.GetMarketingPrivacyUseCase
-import org.sopt.certi.domain.usecase.user.PatchMarketingAgreementUseCase
-import org.sopt.certi.domain.usecase.user.PatchPrivacyAgreementUseCase
 import org.sopt.certi.presentation.ui.setting.sideEffect.SettingSideEffect
 import org.sopt.certi.presentation.ui.setting.state.SettingUiState
 import timber.log.Timber
@@ -21,9 +18,6 @@ import javax.inject.Inject
 
 @HiltViewModel
 class SettingViewModel @Inject constructor(
-    private val getMarketingPrivacyUseCase: GetMarketingPrivacyUseCase,
-    private val patchMarketingAgreementUseCase: PatchMarketingAgreementUseCase,
-    private val patchPrivacyAgreementUseCase: PatchPrivacyAgreementUseCase,
     private val withDrawUseCase: WithDrawUseCase,
     private val tokenManager: TokenManager
 ) : ViewModel() {
@@ -32,90 +26,6 @@ class SettingViewModel @Inject constructor(
 
     private val _sideEffect = Channel<SettingSideEffect>()
     val sideEffect = _sideEffect.receiveAsFlow()
-
-    init {
-        getMarketingPrivacyAgreement()
-    }
-
-    private fun getMarketingPrivacyAgreement() = viewModelScope.launch {
-        getMarketingPrivacyUseCase()
-            .onSuccess { result ->
-                _uiState.update {
-                    it.copy(
-                        switchChecked = result.isAdvertisingAgreement && result.isPrivacyAgreement,
-                        checkboxChecked = result.isPrivacyAgreement
-                    )
-                }
-            }
-            .onFailure { error ->
-                Timber.d(error, "광고성 정보 수신 동의 정보 불러오기 실패")
-            }
-    }
-
-    fun onSwitchCheckChange(checked: Boolean) {
-        if (_uiState.value.checkboxChecked) {
-            viewModelScope.launch {
-                patchMarketingAgreementUseCase(checked)
-                    .onSuccess {
-                        _uiState.update { it.copy(switchChecked = checked) }
-                    }
-                    .onFailure { error ->
-                        Timber.e(error, "광고성 정보 수신 동의 업데이트 실패")
-                    }
-            }
-        } else {
-            _uiState.update { it.copy(isMarketingConfirmDialogVisible = true) }
-        }
-    }
-
-    fun onMarketingConfirmDialogConfirm() = viewModelScope.launch {
-        patchPrivacyAgreementUseCase(true)
-            .onSuccess {
-                patchMarketingAgreementUseCase(true)
-                    .onSuccess {
-                        _uiState.update {
-                            it.copy(
-                                switchChecked = true,
-                                checkboxChecked = true,
-                                isMarketingConfirmDialogVisible = false
-                            )
-                        }
-                        showMarketingSnackbar()
-                    }
-                    .onFailure { error ->
-                        Timber.e(error, "광고성 정보 수신 동의 실패")
-                    }
-            }
-            .onFailure { error ->
-                Timber.e(error, "개인정보 수집 및 이용 동의 실패")
-            }
-    }
-
-    fun onMarketingConfirmDialogDismiss() {
-        _uiState.update { it.copy(isMarketingConfirmDialogVisible = false) }
-    }
-
-    fun onCheckboxCheckChange(checked: Boolean) = viewModelScope.launch {
-        patchPrivacyAgreementUseCase(checked)
-            .onSuccess {
-                _uiState.update { state ->
-                    state.copy(
-                        checkboxChecked = checked,
-                        switchChecked = if (!checked) false else state.switchChecked
-                    )
-                }
-                if (checked) showMarketingSnackbar()
-            }
-            .onFailure { error ->
-                Timber.e(error, "개인정보 동의 업데이트 실패")
-            }
-    }
-
-    private fun showMarketingSnackbar() {
-        viewModelScope.launch {
-            _sideEffect.send(SettingSideEffect.ShowMarketingConfirmSnackbar)
-        }
-    }
 
     fun onLogoutClick() {
         _uiState.update { it.copy(isLogoutDialogVisible = true) }
