+++ app/src/main/java/org/sopt/certi/presentation/ui/setting/sideEffect/SettingNotificationSideEffect.kt
@@ -1,0 +1,5 @@
+package org.sopt.certi.presentation.ui.setting.sideEffect
+
+sealed interface SettingNotificationSideEffect {
+    data object ShowMarketingConfirmSnackbar : SettingNotificationSideEffect
+}
