+++ app/src/main/java/org/sopt/certi/presentation/ui/setting/sideEffect/SettingSideEffect.kt
@@ -1,6 +1,5 @@
 package org.sopt.certi.presentation.ui.setting.sideEffect
 
 sealed interface SettingSideEffect {
-    data object ShowMarketingConfirmSnackbar : SettingSideEffect
     data object NavigateToLogin : SettingSideEffect
 }
