+++ app/src/main/java/org/sopt/certi/presentation/ui/setting/state/SettingUiState.kt
@@ -1,9 +1,6 @@
 package org.sopt.certi.presentation.ui.setting.state
 
 data class SettingUiState(
-    val switchChecked: Boolean = false,
-    val checkboxChecked: Boolean = false,
-    val isMarketingConfirmDialogVisible: Boolean = false,
     val isLogoutDialogVisible: Boolean = false,
     val isDeleteAccountDialogVisible: Boolean = false
 )
